seperate HomSearch file
