+++ src/wiring_diagrams/Directed.jl
@@ -31,7 +31,7 @@ export AbstractBox, Box, WiringDiagram, SchWiringDiagram,
 using StructEquality
 
 using ...GATs, ...Graphs.BasicGraphs, ...CategoricalAlgebra.CSets
-import ...CategoricalAlgebra.CSets: is_isomorphic
+import ...CategoricalAlgebra.HomSearch: is_isomorphic
 import ...CategoricalAlgebra.FinCats: graph
 import ...Graphs: all_neighbors, neighbors, outneighbors, inneighbors
